# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: NODAQ API — AI-powered business management for French SMEs
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /factures:
    get:
      operationId: listFactures
      tags: [factures]
      summary: List invoices
      parameters:
        - in: query
          name: settled
          schema:
            type: boolean
      responses:
        "200":
          description: Invoice list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FactureList"
    post:
      operationId: createFacture
      tags: [factures]
      summary: Create an invoice
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FactureInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Facture"
  /factures/{id}:
    patch:
      operationId: updateFacture
      tags: [factures]
      summary: Update invoice (mark settled, etc.)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FactureUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Facture"
        "404":
          description: Not found

  # ── Prospects ────────────────────────────────────────────────────────────
...
  /devis:
    get:
      operationId: listDevis
      tags: [devis]
      summary: List devis
      parameters:
        - in: query
          name: statut
          schema:
            type: string
        - in: query
          name: search
          schema:
            type: string
      responses:
        "200":
          description: List of devis
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DevisList"
    post:
      operationId: createDevis
      tags: [devis]
      summary: Create a devis
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DevisInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Devis"
  /devis/{id}:
    get:
      operationId: getDevis
      tags: [devis]
      summary: Get a devis
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Devis detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Devis"
        "404":
          description: Not found
    patch:
      operationId: updateDevis
      tags: [devis]
      summary: Update a devis
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DevisUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Devis"
        "404":
          description: Not found
    delete:
      operationId: deleteDevis
      tags: [devis]
      summary: Delete a devis
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
  /devis/{id}/convert:
    post:
      operationId: convertDevisToAffaire
      tags: [devis]
      summary: Convert accepted devis to affaire
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Affaire created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Affaire"
        "404":
          description: Not found

  # ── Classeur ────────────────────────────────────────────────────────────────
...
  /echeances:
    get:
      operationId: listEcheances
      tags: [echeances]
      summary: List fiscal deadlines
      parameters:
        - in: query
          name: statut
          schema:
            type: string
      responses:
        "200":
          description: List of echeances
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Echeance"
    post:
      operationId: createEcheance
      tags: [echeances]
      summary: Create a fiscal deadline
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EcheanceInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Echeance"
  /echeances/{id}:
    patch:
      operationId: updateEcheance
      tags: [echeances]
      summary: Update a fiscal deadline
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EcheanceUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Echeance"
        "404":
          description: Not found
    delete:
      operationId: deleteEcheance
      tags: [echeances]
      summary: Delete a fiscal deadline
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found

  # ── Marge ────────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Affaire:
      type: object
      required: [id, label, status, createdAt, updatedAt]
      properties:
        id:
          type: string
        reference:
          type: ["string", "null"]
        label:
          type: string
        clientName:
          type: ["string", "null"]
        status:
          type: string
          enum: [PROSPECT, DEVIS_ENVOYE, ACCEPTEE, EN_COURS, TERMINEE, PERDUE, ARCHIVEE]
        quotedAmountCents:
          type: ["number", "null"]
        invoicedAmountCents:
          type: ["number", "null"]
        marginCents:
          type: ["number", "null"]
        notes:
          type: ["string", "null"]
        startDate:
          type: ["string", "null"]
          format: date
        completedAt:
          type: ["string", "null"]
          format: date
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    Facture:
      type: object
      required: [id, customerName, number, issuedDate, dueDate, amountCents, settled, createdAt]
      properties:
        id:
          type: string
        customerName:
          type: string
        number:
          type: string
        issuedDate:
          type: string
          format: date
        dueDate:
          type: string
          format: date
        amountCents:
          type: number
        residualCents:
          type: number
        settled:
          type: boolean
        affaireId:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
    FactureList:
      type: object
      required: [factures, totalAmountCents, totalOverdueCents]
      properties:
        factures:
          type: array
          items:
            $ref: "#/components/schemas/Facture"
        totalAmountCents:
          type: number
        totalOverdueCents:
          type: number
    FactureInput:
      type: object
      required: [customerName, number, issuedDate, dueDate, amountCents]
      properties:
        customerName:
          type: string
          minLength: 1
        number:
          type: string
          minLength: 1
        issuedDate:
          type: string
          format: date
        dueDate:
          type: string
          format: date
        amountCents:
          type: number
        affaireId:
          type: string
    FactureUpdate:
      type: object
      properties:
        settled:
          type: boolean
        residualCents:
          type: number
        affaireId:
          type: string

    # ── Prospects ────────────────────────────────────────────────────────────
...
    DevisLine:
      type: object
      required: [id, description, quantity, unitPriceCents]
      properties:
        id:
          type: string
        description:
          type: string
        quantity:
          type: number
        unitPriceCents:
          type: number
    Devis:
      type: object
      required: [id, reference, clientName, status, totalHTCents, totalTTCCents, tvaRate, remise, lines, createdAt, updatedAt]
      properties:
        id:
          type: string
        reference:
          type: string
        clientName:
          type: string
        status:
          type: string
          enum: [BROUILLON, ENVOYE, ACCEPTE, REFUSE, EXPIRE]
        lines:
          type: array
          items:
            $ref: "#/components/schemas/DevisLine"
        totalHTCents:
          type: number
          description: Total HT en centimes (entier)
        totalTTCCents:
          type: number
          description: Total TTC en centimes (entier)
        tvaRate:
          type: number
        remise:
          type: number
        notes:
          type: ["string", "null"]
        validUntil:
          type: ["string", "null"]
          format: date
        affaireId:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    DevisList:
      type: object
      required: [devis, total, totalTTCCents]
      properties:
        devis:
          type: array
          items:
            $ref: "#/components/schemas/Devis"
        total:
          type: number
        totalTTCCents:
          type: number
          description: Somme des totaux TTC en centimes
    DevisInput:
      type: object
      required: [clientName, lines]
      properties:
        clientName:
          type: string
          minLength: 1
        status:
          type: string
        lines:
          type: array
          items:
            $ref: "#/components/schemas/DevisLine"
        tvaRate:
          type: number
        remise:
          type: number
        notes:
          type: string
        validUntil:
          type: string
          format: date
    DevisUpdate:
      type: object
      properties:
        clientName:
          type: string
        status:
          type: string
        lines:
          type: array
          items:
            $ref: "#/components/schemas/DevisLine"
        tvaRate:
          type: number
        remise:
          type: number
        notes:
          type: string
        validUntil:
          type: string
          format: date

    # ── Classeur ──────────────────────────────────────────────────────────────
...
    Echeance:
      type: object
      required: [id, type, label, dueDate, status, createdAt]
      properties:
        id:
          type: string
        type:
          type: string
          enum: [TVA, IS, URSSAF, CFE, CVAE, AUTRE]
        label:
          type: string
        dueDate:
          type: string
          format: date
        estimatedCents:
          type: ["number", "null"]
        paidCents:
          type: ["number", "null"]
        status:
          type: string
          enum: [A_VENIR, PAYEE, EN_RETARD]
        notes:
          type: ["string", "null"]
        paidAt:
          type: ["string", "null"]
          format: date-time
        createdAt:
          type: string
          format: date-time
    EcheanceInput:
      type: object
      required: [type, label, dueDate]
      properties:
        type:
          type: string
        label:
          type: string
          minLength: 1
        dueDate:
          type: string
          format: date
        estimatedCents:
          type: number
        notes:
          type: string
    EcheanceUpdate:
      type: object
      properties:
        label:
          type: string
        dueDate:
          type: string
          format: date
        estimatedCents:
          type: number
        paidCents:
          type: number
        status:
          type: string
        notes:
          type: string

    # ── Marge ─────────────────────────────────────────────────────────────────
...