# lib/api-zod/src/generated/types/devisInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * NODAQ API — AI-powered business management for French SMEs
 * OpenAPI spec version: 0.1.0
 */
import type { DevisLine } from './devisLine';

export interface DevisInput {
  /** @minLength 1 */
  clientName: string;
  status?: string;
  lines: DevisLine[];
  tvaRate?: number;
  remise?: number;
  notes?: string;
  validUntil?: Date;
}