# lib/api-zod/src/generated/types/devisLine.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * NODAQ API — AI-powered business management for French SMEs
 * OpenAPI spec version: 0.1.0
 */

export interface DevisLine {
  id: string;
  description: string;
  quantity: number;
  unitPriceCents: number;
}