# lib/api-zod/src/generated/types/devisList.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * NODAQ API — AI-powered business management for French SMEs
 * OpenAPI spec version: 0.1.0
 */
import type { Devis } from './devis';

export interface DevisList {
  devis: Devis[];
  total: number;
  /** Somme des totaux TTC en centimes */
  totalTTCCents: number;
}